requirements.txt
absl-py==0.9.0
appdirs==1.4.4
astunparse==1.6.3
backcall==0.1.0
bitmap==0.0.7
bkcharts==0.2
bleach==3.1.0
boto==2.49.0
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
configobj==5.0.6
cycler==0.10.0
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
flask_pymongo==2.3.0
future==0.18.2
gast==0.3.3
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
gunicorn==19.3.0
h5py==2.10.0
httplib2==0.9.2
idna==2.6
itsdangerous==0.24
Jinja2==2.11.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Mako==1.0.7
Markdown==3.2.2
MarkupSafe==0.23
matplotlib==3.2.1
mnist==0.2.2
netifaces==0.10.4
numpy==1.18.5
oauthlib==3.1.0
opencv-python-headless==4.2.0.32
opt-einsum==3.2.1
packaging==20.4
pexpect==4.2.1
Pillow==7.1.2
pipenv==2020.6.2
protobuf==3.12.2
psutil==5.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pyinotify==0.9.6
pymacaroons==0.13.0
pymongo==3.11.0
pyparsing==2.4.7
pyRFC3339==1.0
python-dateutil==2.8.1
python-debian==0.1.32
python-xlib==0.20
pytz==2018.3
pyxdg==0.25
reportlab==3.4.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.0
scipy==1.4.1
setproctitle==1.1.10
six==1.15.0
systemd-python==234
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow-cpu==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tinycss==0.4
tornado==6.0.4
urllib3==1.22
virtualenv==20.0.21
virtualenv-clone==0.5.4
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
